requirements.txt
torch>=2.0.0
torchvision>=0.15.0
diffusers>=0.25.0
transformers>=4.35.0
accelerate>=0.25.0
xformers
numpy
pandas
matplotlib
opencv-python
pillow
scipy
scikit-image
tqdm
tensorboard
wandb
lpips
pytorch-fid
pycocotools
detectron2 @ git+https://github.com/facebookresearch/detectron2.git
einops
omegaconf
huggingface_hub
safetensors
